+++ Cloudpos/src/main/java/org/example/cloudpos/product/service/ProductService.java
@@ -1,5 +1,6 @@
-package org.example.cloudpos.product;
+package org.example.cloudpos.product.service;
 
+import org.example.cloudpos.product.domain.ProductStatus;
 import org.example.cloudpos.product.dto.ProductCreateRequest;
 import org.example.cloudpos.product.dto.ProductResponse;
 import org.example.cloudpos.product.dto.ProductUpdateRequest;
